merge

+++ src/main/java/com/ucu/olimpiadas/Model/IDiscipline.java
@@ -1,10 +1,8 @@
 package com.ucu.olimpiadas.Model;
 
-import jakarta.servlet.http.Part;
-
 public interface IDiscipline {
     public String getName();
     public ICategory getCategory();
     public Participant getAthlete();
-    public int calculatePoints();
+    public double calculatePoints();
 }
